Save state, dup'ed as fuck
The code's a fucking mess right now. I copy-pasted between my encrypt
and decrypt functions (I know...) so the next thing I need to do is
write a generic function that encrypts/decrypts based on a mode switch.
Then I'll be able to tackle the multitude of other problems with this
code.