FUCK YEAH
I FINALLY figured out challenge #11, detecting ECB/CBC modes. The first
breakthrough came when I realized I misunderstood the requirements of
the challenge -- I was to append chosen bytes, not random bytes. The
second breakthrough was when I realized/remembered that ECB is
block-based, i.e. a repetition in the text doesn't necessarily produce a
repetition in the ciphertext unless the whole block is repeated. The
third and crucial breakthrough came when I realized my chosen plaintext
had to be 16 bytes long in order to guarantee a repetition consistently
-- it almost seemed too easy, like it was cheating, but I thought, hey,
this is a chosen plaintext attack after all. Looking forward to the next
challenge!